backend/app/import_export.py
"""Módulo de importação e exportação de planilhas Excel.

Implementa funções que fazem a ponte entre arquivos `.xlsx`/`.xlsm` e as
entidades do sistema, de acordo com o mapeamento definido em
`mapeamento.xlsx.yaml`. Para simplificação, as rotinas completas não
estão implementadas, mas o esqueleto ilustra como carregar a planilha
com openpyxl e iterar pelas abas/linhas.
"""

from typing import Tuple
from pathlib import Path
import yaml
from openpyxl import load_workbook

from . import models, rule_engine

ROOT_DIR = Path(__file__).resolve().parents[2]
MAP_PATH = ROOT_DIR / "mapeamento.xlsx.yaml"


def load_mapping() -> dict:
    with open(MAP_PATH, "r", encoding="utf-8") as f:
        return yaml.safe_load(f)


def import_workbook(file_path: Path, db, current_user) -> Tuple[int, int]:
    """Importa avaliações a partir de um arquivo Excel.

    :returns: tupla (importados, atualizados)
    """
    mapping = load_mapping()
    wb = load_workbook(filename=str(file_path), data_only=True)
    imported = 0
    updated = 0
    # TODO: percorrer abas e linhas aplicando mapeamento
    # Para cada linha, localizar ou criar Resultado/Avaliacao e persistir
    return imported, updated


def export_workbook(result_id: int, format: str = "xlsx", db=None) -> bytes:
    """Exporta a avaliação de um resultado para um arquivo Excel ou DOCX.

    :returns: bytes do arquivo gerado
    """
    # TODO: gerar workbook com openpyxl ou DOCX com python-docx
    raise NotImplementedError