backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.30
pydantic==1.10.14
python-jose==3.3.0
passlib[bcrypt]==1.7.4
openpyxl==3.1.2
PyYAML==6.0.1
python-docx==1.1.2
psycopg2-binary==2.9.9
pytest==7.4.4
pytest-asyncio==0.23.5